+++ core/test/ScanProps.hs
@@ -89,7 +89,7 @@ boundaries =
     && map (gradeWith (0, 0)) [0, 1] == [0, 1]
 
 wireReq :: [[Integer]] -> Value
-wireReq = rowsRequest "6.4.0" "scan.request"
+wireReq = rowsRequest "6.5.0" "scan.request"
 
 replyObj :: Value -> Maybe Object
 replyObj = replyObjWith respond